Stage 4

+++ src/engine/domain/Quiz.kt
@@ -1,4 +1,4 @@
 package engine.domain
 
-data class Quiz(val id: Int, val title: String, val text: String, val options: List<String>, val answer: List<Int>)
+data class Quiz(val id: Long, val title: String, val text: String, val options: List<String>, val answer: List<Int>)
