+++ src/engine/dto/QuizResponse.kt
@@ -1,3 +1,3 @@
 package engine.dto
 
-data class QuizResponse(val id: Int, val title: String, val text: String, val options: List<String>)
+data class QuizResponse(val id: Long, val title: String, val text: String, val options: List<String>)
